Fix delete_object and write convergence in transaction.
Fix a bug, when delete_object was deleting the record if it was written
in the same transaction even if it was written to a different value.

To verify:
%% Create a set table
mnesia:create_table(foo, []).

%% Write and delete_object in transaction
mnesia:transaction(fun() ->
    mnesia:write({foo, bar, one}),
    mnesia:delete_object({foo, bar, not_one})
end).

{atomic, [{foo, bar, one}]} = mnesia:transaction(fun() -> mnesia:read(foo, bar) end).

Added a section to isolation tests to check for non-matching delete_object
requests.