Include question mark in query string on redirect
According to the code in httpc.erl, the pquery parameter in the
#request record should start with a question mark. However, when
a server returned a 403 response, the Location header was parsed
and the question mark was not added to the beginning of the query
string. This ultimately causes the redirect to fail, as instead of
redirecting to "/path?query", httpc redirected to "/pathquery".