[socket-doc-nif] Preliminary socket option table
Added a commented out (preliminary) table in the getopt function
for the socket level. The table is not included (atleast) in the
man page, so there is no point in including it yet. But since its
table(s) of all socket options (one for each level) we need
comprehensive tables "somewhere", but where...

OTP-14831