[socket-nif] Add (partial) support for socket (level sctp) option associnfo
Added support for the SCTP option ASSOCINFO.

This option is a bit tricky. As the underlying structure (sctp_assocparams)
contains the assoc_id, it begs the question what happens if this option
is fetched for:

   * The own assoc (which means that we might have the assoc id in the
     descriptor and can initiate that part of the struct accordningly).
   * Another assoc: From assoc A asks for info with assoc_id set to
     that of assoc B.
   * The "owning" endpoint.
   * Another endpoint (an endpoint to which the assoc does not belong).

So, if the user calls socket:[getopt|setopt] for an association socket,
shall we require that the assoc_id field is set to -1? Or not set at all
and therefor filled in automatically by the nif-code?

And, if the user calls socket:[getopt|setopt] for an endpoint socket,
shall we require that the assoc_id field is set to a valid id? Or shall
it not be allowed?

Questions, questions...

OTP-14831