[socket-nif] Add (partial) support for socket (level sctp) option rtoinfo
Added support for the SCTP option RTOINFO.

We have the same questions for this option as for ASSOCINFO.
Maybe the assoc field shall be made 'out' only (that is, it will
be ignored for setopt). The assoc id used will be that which
is stored in the descriptor (how do we get it to begin with?).

Questions, questions...

OTP-14831