[socket-nif] Add support for socket (level ip) option freebind
Added support for ip level socket option FREEBIND.

Note that there is an option available on FreeBSD called
IP_BINDANY, which seems to have similar properties (FREEBIND
is *not* available on FreeBSD). There are some restrictions
for this option though (which is not mentioned in the Linux
man page).

OTP-14831