[socket-nif] Add support for socket (level ipv6) option dstopts
Added support for the IPv6 socket option(s) DSTOPTS.

The option returns with einval when calling setsockopt,
so either you need to be a privileged user to update, or its not
actually possible to update this option (even though it says
nothing about that in the man page. It only talks about set).
This is the same behaviour as with RTHDR and HOPLIMIT.
On FreeBSD, it says that HOPOPTS requires superuser privileges.
Needs furher checking.

OTP-14831.