[socket-nif] Add support for socket (level ipv6) option flowinfo
Added support for the IPv6 socket option(s) FLOWINFO.

The option returns with einval when calling setsockopt,
so either you need to be a privileged user to update, or its not
actually possible to update this option (even though it says
nothing about that in the man page. It only talks about set).
This is the same behaviour as with DSTOPTS.
Needs furher checking.

OTP-14831.