Do not allocate a new map when the value is the same
This patch optimizes map operations to not allocate new maps
when the key is being replaced by the exact same value in memory.

Imagine this very common idiom:

    Map#{key := compute_new_value(Value, Condition)}

where:

    compute_new_x(X, true) -> X + 1;
    compute_new_x(X, false) -> X;

In many cases, we are not changing the value in `Key`, however
the code prior to this patch would still allocate a new array
for the map values. This optimization changes this.

The cost of optimization is minimum, as in the worst case scenario
it only adds a pointer comparison and boolean check. The major benefit
is reducing the GC pressure by avoiding allocating data.

Next we list the operations we have changed alongside the benchmark
results. The benchmarks basically create a map and perform the same
operations, roughly 20000 times, once replacing the key with the same
value, and another with a different value.

* Map#{Key := Value}

For a map with 4 keys, replacing the fourth key 20000 times went from
718us to 539us.

For a map with 8 keys, replacing the fourth key 20000 times went from
976us to 555us.

* maps:update/3

For a map with 4 keys, replacing the fourth key 20000 times went from
673us to 575us.

For a map with 8 keys, replacing the fourth key 20000 times went from
827us to 585us.

* maps:put/3

For a map with 4 keys, replacing the fourth key 20000 times went from
763us to 553us.

For a map with 8 keys, replacing the fourth key 20000 times went from
788us to 561us.

Note that we have ported some optimizations found in maps:update/3
to maps:put/3 while creating this patch.