Do not allocate a new tuple when the value is the same
This patch optimizes tuple operations to not allocate new tuple
when an element is being replaced by the exact same value in memory.

Imagine this very common idiom:

    Record#my_record{key = compute_new_value(Value, Condition)}

where:

    compute_new_x(X, true) -> X + 1;
    compute_new_x(X, false) -> X;

In many cases, we are not changing the value in `key`, however
the code prior to this patch would still allocate a new tuple.
This optimization changes this.

The cost of optimization is minimum, as it only adds a C array
access plus a pointer comparison. The major benefit is reducing
the GC pressure by avoiding allocating data.

We have only changed erlang:setelement/3. The benchmarks basically
create a tuple and perform the same operations, roughly 20000 times,
once replacing the key with the same value, and another with a
different value.

For a tuple with 4 elements, replacing the fourth element 20000 times
went from 557us to 388us.

For a tuple with 8 elements, replacing the fourth element 20000 times
went from 641us to 421us.