Omit include path debug info for +deterministic builds
Compiling the same file with different include paths resulted in
different files with the `+deterministic` flag even if everything
but the paths were identical. This was caused by the absolute path
of each include directory being unconditionally included in a
debug information chunk.

This commit fixes this by only including this information in
non-deterministic builds.