Fix bug in binary matching
The compiler generates incorrect code for the following example:

    decode_binary(_, <<Length, Data/binary>>) ->
      case {Length, Data} of
        {0, _} ->
          %% When converting the match context back to a binary,
          %% Data will be set to the entire original binary,
          %% that is, to <<0>> instead of <<>>.
          {{0, 0, 0}, Data};
        {4, <<Y:16/little, M, D, Rest/binary>>} ->
          {{Y, M, D}, Rest}
      end.

The problem is the delayed sub binary creation optimization, which
is not safe to do in this case.

This commit introduces a heuristic that will disable the delayed
sub binary creation optimization for this example.  Unfortunately, the
heuristic may turn off the optimization when it would actually be
safe. In the OTP codebase, the optimization is turned off in two
instances, once in string.erl and once in dets_v9.erl.

https://bugs.erlang.org/browse/ERL-689