Add include_ellipsis option to ASN.1 compiler
This option is currently only implemented for BER. It is ignored for
other encodings.

ASN.1 modules compiled with this flag will have an '$ellipsis'
elemented added to the record deinition for any extensible SET or
SEQUENCE.

When decoding, any extension elements will be added to the '$ellipsis'
element or key in the form returned by asn1rt_nif:decode_ber_tlv/1.

When encoding '$ellipsis' is ignored.