Support emulator restart with different flags
This change introduces `init:restart/1`, a variant of `init:restart/0`
which takes a list of flags which will override/extend the ones provided
during first boot, to be applied post-restart.

The purpose of this change is to support the ability to partially
boot a system, perform some reconfiguration, and then restart into that
new configuration cleanly. For example:

* Boot into a minimal system using an initial boot script and base config
file
* During boot, an instruction to apply a function is executed, this
function looks up some configuration from the system environment and
fetches secrets from a configuration management system; this
configuration is merged over the base config, and persisted to a new
`sys.config`
* `init:restart/1` is invoked, updating `-boot`, `-config`, and
potentially `-name` and `-setcookie` flags
* The system boots fully configured

A key point here is that this process allows the reconfiguration step to
configure _any_ application, including `kernel`.

This use case is actually already part of OTP, albeit somewhat hidden. Installing
a new release and making it permanent causes `init` to replace the
`-boot` and `-config` flags, so that restarting the emulator will
restart into the newly installed release, and not the old. The addition
of `init:restart/1` is a generalied extension of this to support additional use
cases.