beam_ssa_opt: Add simplification of switch lists
When the argument for a #b_switch{} comes from a phi node
with only literal values, the switch list could be pruned to
only contain the possible values. It could also be possible
to eliminate the failure label.

Also simplify a switch with a single value list or switch that can be
replaced with an is_boolean test.