beam_peep: Add is_boolean optimization of select_val
A select_val instruction that test whether a register is a boolean
like this:

     {select_val,Reg,{f,Fail},{list,[{atom,true},Lbl,{atom,false},Lbl]}}.

can be replaced with an is_boolean test:

     {test,is_boolean,{f,Fail},[Reg]}.
     {jump,{f,Lbl}}.

This optimization is currently done in beam_dead. However, if done in
the beam_peep, it can catch more opportunities to do the optimization,
because after having run beam_jump, labels that were different have
been coalesced.