beam_ssa_pre_codegen: Fix bug in receive fixing
When creating a phi node for the common exit block of a receive,
the code failed to take into account that there could be more
than one predecessor to the exit block for each remove_message.

Rename exit_predessor/3 to exit_predessors/3 and make it return a
list of the predecessors.