ssl: Fix cipher suite handling
Implementations of TLS 1.3 which choose to support prior versions of
TLS SHOULD support TLS 1.2. That is, a TLS 1.3 ClientHello shall
advertise support for TLS 1.2 ciphers in order to be able to connect
to TLS 1.2 servers.

This commit changes the list of the advertised cipher suites to
include old TLS 1.2 ciphers.

Change-Id: Iaece3ac4b66a59dfbe97068b682d6010d74522b8