Avoid object code terminology on general docs/messages
Although the object code terminology is introduced in the code
and debugger modules, the terminology is mostly unknown to
developers unless they have interacted with those modules
directly.

Therefore when reading the docs and stumbling against object
code for the first time, the term is often puzzling. This is
particularly troublesome in case insensitive filesystems, where
attempting to invoke a module with the wrong case leads to
confusing error messages:

    1> erlpress_core:foo().
    beam/beam_load.c(1428): Error loading module 'erlpress_core':
      module name in object code is erlPress_core

    Loading of erlPress_core.beam failed: :badfile

This commit restricts the object code terminology to the
low-level code and its APIs, using "module bytecode"
whenever appropriate.