beam_ssa_type: Remove clause in arith_op_types/2 that can't match
Remove the following clause from the `fun` clauses in arith_op_types/2
because it cannot possibly match:

    (_, any) -> number;

Here is why it cannot match:

The second argument is the accumulator for lists:foldl/3.
Its initial value is `unknown`.

None of the clauses will update the accumulator to `any`, including
the clause that matches `any` in the first argument -- it will set the
accumulator to `number`.

Thus, the accumulator (second argument) can never be `any` and
the clause can never match. QED.