Move allocation combining from beam_flatten to beam_ssa_codegen
Continuing the simplification of beam_flatten, move the optimization
that eliminates a test_heap instruction following a binary construction
by incorporating the allocation of the heap space into the bs_init*
instruction itself.

This change does not change the generated code in any way.

Also remove beam_utils:combine_heap_needs/2, because beam_flatten
was the last user of it.