[socket-nif|doc] Cleanup and fixed setopt doc
Removed some cruft. Also tried, without success, to fix the
build issue ("Error file: cannot find module exporting type").
It may be because of a spec-file issue.
Finally fixed the setopt doc. It did not have a seperate clause
for clause 8 (the fallback clause that takes level and key as
integer).