[socket-nif|test] Updated test case for connect timeout
The connect timeout test case uses the listen backlog argument
to test connect timeout (its set to 1). The idea is that when
the server backlog is full, it should not reply. The client attempting
to connect will then get a timeout.
But as it turns out, this behaviour is not "set in stone". On FreeBSD
11.2, the result is instead a 'econnreset'.

It should also be noted that this local connections. That is, both end
of the connection is on the same machine.

So, test case has been updated to also accept a econnreset (even
though that is not what it is testing).

OTP-14831